requirements.txt
tidalapi>=0.8.11
textual>=8.0.0
textual-image>=0.8.5
Pillow>=11.0.0
